Fix Room KSP TokenUsageConverter type resolution
